Whole service was rewritten from MongoDB to PostgreSQL Authorization check for update/delete recipe endpoints has been added Forbidden exception has been added

+++ recipes/src/main/java/com/markiewicz/recipes/recipe/RecipeService.java
@@ -1,40 +1,51 @@
 package com.markiewicz.recipes.recipe;
 
+import com.markiewicz.recipes.exception.ForbiddenException;
 import com.markiewicz.recipes.exception.RecipeNotFoundException;
+
+import com.markiewicz.recipes.user.UserService;
 import lombok.AllArgsConstructor;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
-@Service
 @AllArgsConstructor
+@Service
 public class RecipeService {
-    RecipeRepository recipeRepository;
+    private final RecipeRepository recipeRepository;
+    UserService userService;
 
     public List<Recipe> getAllRecipes() {
         return recipeRepository.findAll();
     }
 
-    public Recipe getRecipeById(String id) {
+    public Recipe getRecipeById(Long id) {
         return recipeRepository
                 .findById(id)
                 .orElseThrow(RecipeNotFoundException::new);
     }
 
     public Recipe addNewRecipe(Recipe recipe) {
-        return recipeRepository.insert(recipe);
+        recipe.setDate(LocalDateTime.now());
+        recipe.setUser(userService.getLoggedInUser());
+        return recipeRepository.save(recipe);
     }
 
     @Transactional
-    public void updateRecipe(String id, Recipe recipe) {
+    public void updateRecipe(Long id, Recipe recipe) {
         Recipe updatedRecipe = recipeRepository
                 .findById(id)
                 .orElseThrow(RecipeNotFoundException::new);
 
-        updateRecipeFields(updatedRecipe, recipe);
+        if(checkCompatibility(updatedRecipe)) {
+            updateRecipeFields(updatedRecipe, recipe);
+        } else {
+            throw new ForbiddenException();
+        }
     }
 
     private void updateRecipeFields(Recipe updatedRecipe, Recipe recipe) {
@@ -46,11 +57,21 @@ public class RecipeService {
         updatedRecipe.setDate(LocalDateTime.now());
     }
 
-    public void removeRecipeById(String id) {
+    public void removeRecipeById(Long id) {
         Recipe recipe = recipeRepository
                 .findById(id)
                 .orElseThrow(RecipeNotFoundException::new);
 
-        recipeRepository.delete(recipe);
+        if (checkCompatibility(recipe)) {
+            recipeRepository.delete(recipe);
+        } else {
+            throw new ForbiddenException();
+        }
+    }
+
+    private boolean checkCompatibility(Recipe recipe) {
+        Long loggedInUserId = userService.getLoggedInUser().getId();
+        Long recipeAuthorId = recipe.getUser().getId();
+        return Objects.equals(loggedInUserId, recipeAuthorId);
     }
 }
